+++ frontend/src/components/NewSubscriptionPlanForm/styles.scss
@@ -1,0 +1,18 @@
+#new-subscription-plan-form{
+
+    background-color:  #FFF8E4;
+    border-radius: 15px;
+    padding-bottom: 1rem;
+    
+    
+
+    .bttn{
+        margin-top: 1rem;
+    }
+
+    .submit{
+        width: 10rem;
+        border-radius: 20px;
+        background-color: #C78A61;
+    }
+}
