--- frontend/src/routes/NewSubscriptionPlan/styles.scss
@@ -1,18 +1,0 @@
-#new-subscription-plan{
-
-    background-color:  #FFF8E4;
-    border-radius: 15px;
-    padding-bottom: 1rem;
-    
-    
-
-    .bttn{
-        margin-top: 1rem;
-    }
-
-    .submit{
-        width: 10rem;
-        border-radius: 20px;
-        background-color: #C78A61;
-    }
-}
